Fix test coverage on core
